langchain/ragname_deepseek/requirements.txt
langchain==0.3.19
langchain_community==0.3.18
langchain_deepseek==0.1.2
python-dotenv==1.0.1
